new: allow player to pick up gizmo which the camera follows
Does this by having a Marker2D on the player to mark a place for the marker to rest, and an Area2D on the gizmo to act as a trigger to mark that the player is in range to pick it up

picking up the gizmo disables collision.

careful when enabling position smoothing--it kinda freaks out. I assume we don\'t want that anyways.